+++ src/components/App/contexts/CombinedContextProvider.jsx
@@ -1,0 +1,14 @@
+import { UserDataContext } from "./UserDataProvider";
+import { ProfilePictureContext } from "./ProfilePictureProvider";
+
+function CombinedContextProvider({children}) {
+    return (
+        <UserDataContext.Consumer>
+            <ProfilePictureContext.Consumer>
+                {children}
+            </ProfilePictureContext.Consumer>
+        </UserDataContext.Consumer>
+    )
+}
+
+export default CombinedContextProvider
