+++ src/hooks/handleSignOut.js
@@ -1,0 +1,6 @@
+function handleSignOut() {
+    sessionStorage.removeItem("token");
+    setToken(false);
+}
+
+export default handleSignOut()
